# sira-backend-master 2/swagger-api.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/units/{id}": {
            "get": {
                "operationId": "UnitsController_findOne",
                "summary": "Get a unit by ID (sales: own or published only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "UnitsController_update",
                "summary": "Update unit inventory (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UnitsController_remove",
                "summary": "Delete a unit (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/unit-previews": {
            "post": {
                "operationId": "UnitPreviewsController_create",
                "summary": "Request a preview — own unit = scheduled, other = pending",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePreviewDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "UnitPreviewsController_findAll",
                "summary": "List previews (filters: unitId, status, requestedById)",
                "parameters": [
                    {
                        "name": "unitId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "requestedById",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salesId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "teamId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/my-units": {
            "get": {
                "operationId": "UnitPreviewsController_findForMyUnits",
                "summary": "Previews on units I own (for approval)",
                "parameters": [
                    {
                        "name": "q",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salesId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "teamId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/availability": {
            "get": {
                "operationId": "UnitPreviewsController_checkAvailability",
                "summary": "Check if a unit is free for a preview window (overlapping pending/scheduled/checked-in)",
                "parameters": [
                    {
                        "name": "unitId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "scheduledAt",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "durationMin",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excludePreviewId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}": {
            "get": {
                "operationId": "UnitPreviewsController_findOne",
                "summary": "Get a single preview (requester, unit owner, or lead roles)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}/approve": {
            "patch": {
                "operationId": "UnitPreviewsController_approve",
                "summary": "Approve a pending preview (unit owner only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}/reject": {
            "patch": {
                "operationId": "UnitPreviewsController_reject",
                "summary": "Reject a pending preview (unit owner only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}/check-in": {
            "patch": {
                "operationId": "UnitPreviewsController_checkIn",
                "summary": "Check in with GPS location",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckInDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}/check-out": {
            "patch": {
                "operationId": "UnitPreviewsController_checkOut",
                "summary": "Check out with GPS location",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckOutDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/unit-previews/{id}/cancel": {
            "patch": {
                "operationId": "UnitPreviewsController_cancel",
                "summary": "Cancel a preview (requester only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Unit Previews"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "CRM API",
        "description": "Comprehensive CRM API for real estate management",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "UpdateUnitDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Omit to auto-generate a unique code (recommended)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "description": "Short unit / listing address line"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "monthlyInstallment": {
                        "type": "number",
                        "description": "Monthly installment"
                    },
                    "deliveryDate": {
                        "type": "string",
                        "description": "Delivery / handover date (YYYY-MM-DD)"
                    },
                    "bedrooms": {
                        "type": "number"
                    },
                    "bathrooms": {
                        "type": "number"
                    },
                    "floor": {
                        "type": "number",
                        "description": "Floor number (building level)"
                    },
                    "area": {
                        "type": "number",
                        "description": "Area in sqm or sqft"
                    },
                    "unitType": {
                        "type": "string"
                    },
                    "driveMediaLink": {
                        "type": "string",
                        "description": "Google Drive (or similar) link to media / documents — paste share link",
                        "maxLength": 1024
                    },
                    "amenities": {
                        "description": "Amenity labels (e.g. tags)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "externalLinks": {
                        "description": "External URLs (listings, walkthroughs, etc.)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "reserved",
                            "sold",
                            "unavailable"
                        ],
                        "default": "available"
                    }
                }
            },
...
            "CreatePreviewDto": {
                "type": "object",
                "properties": {
                    "unitId": {
                        "type": "number",
                        "description": "Unit ID to preview"
                    },
                    "clientName": {
                        "type": "string",
                        "description": "Client full name"
                    },
                    "clientPhone": {
                        "type": "string",
                        "description": "Client phone number"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "description": "Scheduled date-time (ISO 8601)"
                    },
                    "durationMin": {
                        "type": "number",
                        "description": "Duration in minutes (default 60)"
                    },
                    "notes": {
                        "type": "string",
                        "description": "Notes about the visit"
                    }
                },
                "required": [
                    "unitId",
                    "clientName",
                    "scheduledAt"
                ]
            },
            "CheckInDto": {
                "type": "object",
                "properties": {
                    "lat": {
                        "type": "number",
                        "description": "Latitude of check-in location"
                    },
                    "lng": {
                        "type": "number",
                        "description": "Longitude of check-in location"
                    }
                }
            },
            "CheckOutDto": {
                "type": "object",
                "properties": {
                    "lat": {
                        "type": "number",
                        "description": "Latitude of check-out location"
                    },
                    "lng": {
                        "type": "number",
                        "description": "Longitude of check-out location"
                    }
                }
            },
...
        }
    }
}